fix:  about pmx compilation error & element_array

+++ libMMD/source/assets/public/file_element.hpp
@@ -117,7 +117,7 @@ namespace libmmd
 			return true;
 		}
 
-		int32_t size() const
+		uint64_t size() const
 		{
 			return data_.size();
 		}
