+++ libMMD/source/assets/public/vmd_element.hpp
@@ -115,7 +115,7 @@ namespace libmmd
 		vmd_element_array_impl& operator =(vmd_element_array_impl&&) noexcept = default;
 
 
-		[[nodiscard]] int32_t size() const override
+		[[nodiscard]] uint64_t size() const override
 		{
 			return array_.size();
 		}
